+++ src/cli.rs
@@ -5,6 +5,7 @@ pub use self::command::Command;
 
 use clap::{App, Arg, ArgMatches};
 
+use self::commands::MigrateCommand;
 use self::commands::ProjectCommand;
 use crate::config::Config;
 
@@ -20,11 +21,13 @@ pub fn app() -> App<'static, 'static> {
                 .takes_value(true),
         )
         .subcommand(ProjectCommand::cli())
+        .subcommand(MigrateCommand::cli())
 }
 
 pub fn exec(config: &mut Config, args: &ArgMatches<'_>) -> Result<(), &'static str> {
     match args.subcommand() {
         ("project", Some(args)) => ProjectCommand::exec(config, args),
+        ("migrate", Some(args)) => MigrateCommand::exec(config, args),
         ("", None) => Ok(()),
         _ => unreachable!(),
     }