+++ src/cli/commands.rs
@@ -1,3 +1,5 @@
+mod migrate;
 mod project;
 
+pub use self::migrate::MigrateCommand;
 pub use self::project::ProjectCommand;