+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `project list` command
+- Added `migrate` command
 
 ## [0.1.0] - 2018-12-20